Add useGetAuthorizedAPIList call
